pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "restrack"
description = "Python application to for results tracking."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "vvcb" }]
dependencies = [
    "holoviews",
    "pandas",
    "panel",
    "pyodbc",
    "python-dotenv",
    "sqlmodel",
    "watchfiles",
    "fastapi[standard]",
]
dynamic = ["version"]

[tool.setuptools]
packages = ["restrack"]

[tool.setuptools.dynamic]
version = { attr = "restrack.__version__" }

[project.urls]
Homepage = "https://github.com/vvcb/restrack"
Documentation = "https://vvcb.github.io/restrack/"

[dependency-groups]
dev = ["black>=24.10.0", "jupyterlab>=4.3.3", "pre-commit>=4.0.1"]
